tools/tokenserver/requirements.txt
boto==2.49.0
hawkauthlib==2.0.0
mysqlclient==2.1.1
pyramid==2.0.1
sqlalchemy==1.4.46
testfixtures
tokenlib==2.0.0
PyBrowserID==0.14.0
datadog
backoff

